Revolutio/citus/python-client/requirements-spark.txt
azure-identity
azure-storage-blob
kafka-python
numpy
pandas
pyarrow
pyspark
python-dotenv
sseclient
